Make category API endpoints shallow

+++ app/controllers/v1/categories_controller.rb
@@ -1,9 +1,8 @@
 class V1::CategoriesController < ApplicationController
-  before_action :authorize_budget_sheet, only: [:create]
-  before_action :authorize_category, only: [:update, :destroy]
-
   def create
     category = Category.new category_params
+
+    authorize(category)
 
     if category.save
       render json: category, status: 201
@@ -13,6 +12,8 @@ class V1::CategoriesController < ApplicationController
   end
 
   def update
+    authorize(category)
+
     if category.update category_params
       render json: category
     else
@@ -21,15 +22,13 @@ class V1::CategoriesController < ApplicationController
   end
 
   def destroy
+    authorize(category)
+
     category.destroy
     render nothing: true, status: 204
   end
 
   private
-
-  def budget_sheet
-    @budget_sheet ||= BudgetSheet.where(id: budget_sheet_id).first_or_initialize
-  end
 
   def category
     @category ||= Category.find params[:id]
@@ -37,17 +36,5 @@ class V1::CategoriesController < ApplicationController
 
   def category_params
     params.require(:category).permit(:name, :budget_amount, :budget_sheet_id)
-  end
-
-  def budget_sheet_id
-    params.require(:category)[:budget_sheet_id]
-  end
-
-  def authorize_budget_sheet
-    authorize budget_sheet
-  end
-
-  def authorize_category
-    authorize category
   end
 end
